Comprehensions returning objects.
This allows short form list/array/range comprehensions to construct
return objects instead of return arrays. The behaviour is triggered by
delivering a (keyExpression,valueExpression) pair instead of just an
expression. For example:

result = (b,a) for a,b of someObject

..swaps keys with values for someObject. Or to quickly populate an
object one could:

topicSums = (topic,0) for topic in topics